requirements.txt
requests
pandas
numpy
pyyaml